requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
langchain-google-genai
google-generativeai
python-dotenv
pydantic